+++ returns/context/requires_context_future_result.py
@@ -740,37 +740,6 @@ class RequiresContextFutureResult(
             lambda deps: self(deps).bind_async(function),
         )
 
-    def fix(
-        self, function: Callable[[_ErrorType], _NewValueType],
-    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
-        """
-        Composes failed container with a pure function.
-
-        .. code:: python
-
-          >>> import anyio
-          >>> from returns.context import RequiresContextFutureResult
-          >>> from returns.io import IOSuccess
-
-          >>> assert anyio.run(
-          ...     RequiresContextFutureResult.from_value(1).fix(
-          ...        lambda x: x + 1,
-          ...     ),
-          ...     RequiresContextFutureResult.empty,
-          ... ) == IOSuccess(1)
-
-          >>> assert anyio.run(
-          ...     RequiresContextFutureResult.from_failure(1).fix(
-          ...        lambda x: x + 1,
-          ...     ),
-          ...     RequiresContextFutureResult.empty,
-          ... ) == IOSuccess(2)
-
-        """
-        return RequiresContextFutureResult(
-            lambda deps: self(deps).fix(function),
-        )
-
     def alt(
         self, function: Callable[[_ErrorType], _NewErrorType],
     ) -> 'RequiresContextFutureResult[_ValueType, _NewErrorType, _EnvType]':