+++ returns/context/requires_context_ioresult.py
@@ -455,28 +455,6 @@ class RequiresContextIOResult(
             lambda deps: self(deps).bind(function),
         )
 
-    def fix(
-        self, function: Callable[[_ErrorType], _NewValueType],
-    ) -> 'RequiresContextIOResult[_NewValueType, _ErrorType, _EnvType]':
-        """
-        Composes failed container with a pure function.
-
-        .. code:: python
-
-          >>> from returns.context import RequiresContextIOResult
-          >>> from returns.io import IOSuccess
-
-          >>> assert RequiresContextIOResult.from_value(1).fix(
-          ...     lambda x: x + 1,
-          ... )(...) == IOSuccess(1)
-
-          >>> assert RequiresContextIOResult.from_failure(1).fix(
-          ...     lambda x: x + 1,
-          ... )(...) == IOSuccess(2)
-
-        """
-        return RequiresContextIOResult(lambda deps: self(deps).fix(function))
-
     def alt(
         self, function: Callable[[_ErrorType], _NewErrorType],
     ) -> 'RequiresContextIOResult[_ValueType, _NewErrorType, _EnvType]':