+++ returns/functions.py
@@ -129,7 +129,7 @@ def raise_exception(exception: Exception) -> NoReturn:
       >>> user: Result[int, ValueError] = Failure(ValueError('boom'))
 
       >>> # Here we unwrap internal exception and raise it:
-      >>> user.fix(raise_exception)
+      >>> user.alt(raise_exception)
       Traceback (most recent call last):
         ...
       ValueError: boom