+++ tests/test_result/test_result_map.py
@@ -12,12 +12,6 @@ def test_alt_failure():
     assert Failure(5).alt(str) == Failure('5')
 
 
-def test_fix_success():
-    """Ensures that Success.fix is NoOp."""
-    assert Success(5).fix(str) == Success(5)
+def test_alt_success():
+    """Ensures that Success.alt is NoOp."""
     assert Success(5).alt(str) == Success(5)
-
-
-def test_fix_failure():
-    """Ensures that Failure.fix produces the Success."""
-    assert Failure(5).fix(str) == Success('5')