+++ returns/context/requires_context_result.py
@@ -14,7 +14,6 @@ from typing_extensions import final
 
 from returns._generated.iterable import iterable
 from returns.context import NoDeps
-from returns.interfaces import applicative, bindable, mappable, rescuable
 from returns.interfaces.specific import result
 from returns.primitives.container import BaseContainer
 from returns.primitives.hkt import Kind3, dekind
@@ -41,10 +40,6 @@ _FirstType = TypeVar('_FirstType')
 class RequiresContextResult(
     BaseContainer,
     Kind3['RequiresContextResult', _ValueType, _ErrorType, _EnvType],
-    mappable.Mappable3[_ValueType, _ErrorType, _EnvType],
-    bindable.Bindable3[_ValueType, _ErrorType, _EnvType],
-    applicative.Applicative3[_ValueType, _ErrorType, _EnvType],
-    rescuable.Rescuable3[_ValueType, _ErrorType, _EnvType],
     result.ResultBased3[_ValueType, _ErrorType, _EnvType],
 ):
     """