+++ returns/methods/__init__.py
@@ -1,6 +1,7 @@
 from returns.methods.apply import apply as apply
 from returns.methods.bind import bind as bind
 from returns.methods.bind_io import bind_io as bind_io
+from returns.methods.bind_ioresult import bind_ioresult as bind_ioresult
 from returns.methods.bind_result import bind_result as bind_result
 from returns.methods.map import map_ as map_
 from returns.methods.rescue import rescue as rescue