+++ returns/pointfree/__init__.py
@@ -1,6 +1,7 @@
 from returns.pointfree.apply import apply as apply
 from returns.pointfree.bind import bind as bind
 from returns.pointfree.bind_io import bind_io as bind_io
+from returns.pointfree.bind_ioresult import bind_ioresult as bind_ioresult
 from returns.pointfree.bind_result import bind_result as bind_result
 from returns.pointfree.map import map_ as map_
 from returns.pointfree.rescue import rescue as rescue